requirements.txt
Flask==3.1.1
waitress==3.0.0
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<2.0
